refactor: auto utils get their own package
