+++ src/main/java/org/team1540/robot2024/RobotContainer.java
@@ -34,8 +34,8 @@ import org.team1540.robot2024.subsystems.vision.AprilTagVision;
 import org.team1540.robot2024.subsystems.vision.AprilTagVisionIO;
 import org.team1540.robot2024.subsystems.vision.AprilTagVisionIOLimelight;
 import org.team1540.robot2024.subsystems.vision.AprilTagVisionIOSim;
-import org.team1540.robot2024.util.AutoCommand;
-import org.team1540.robot2024.util.AutoManager;
+import org.team1540.robot2024.util.auto.AutoCommand;
+import org.team1540.robot2024.util.auto.AutoManager;
 import org.team1540.robot2024.util.PhoenixTimeSyncSignalRefresher;
 import org.team1540.robot2024.util.swerve.SwerveFactory;
 import org.team1540.robot2024.util.vision.VisionPoseAcceptor;
