+++ src/main/java/org/team1540/robot2024/commands/autos/AmpLanePABCSprint.java
@@ -3,8 +3,8 @@ package org.team1540.robot2024.commands.autos;
 import org.team1540.robot2024.subsystems.drive.Drivetrain;
 import org.team1540.robot2024.subsystems.indexer.Indexer;
 import org.team1540.robot2024.subsystems.shooter.Shooter;
-import org.team1540.robot2024.util.AutoCommand;
-import org.team1540.robot2024.util.PathHelper;
+import org.team1540.robot2024.util.auto.AutoCommand;
+import org.team1540.robot2024.util.auto.PathHelper;
 
 public class AmpLanePABCSprint extends AutoCommand {
     public AmpLanePABCSprint(Drivetrain drivetrain, Shooter shooter, Indexer indexer){
