+++ src/main/java/org/team1540/robot2024/commands/autos/SourceLanePHGFSprint.java
@@ -2,8 +2,8 @@ package org.team1540.robot2024.commands.autos;
 
 import edu.wpi.first.wpilibj2.command.PrintCommand;
 import org.team1540.robot2024.subsystems.drive.Drivetrain;
-import org.team1540.robot2024.util.AutoCommand;
-import org.team1540.robot2024.util.PathHelper;
+import org.team1540.robot2024.util.auto.AutoCommand;
+import org.team1540.robot2024.util.auto.PathHelper;
 
 public class SourceLanePHGFSprint extends AutoCommand {
     public SourceLanePHGFSprint(Drivetrain drivetrain){
