+++ src/main/java/org/team1540/robot2024/subsystems/drive/Drivetrain.java
@@ -19,7 +19,7 @@ import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import org.littletonrobotics.junction.AutoLogOutput;
 import org.littletonrobotics.junction.Logger;
-import org.team1540.robot2024.util.LocalADStarAK;
+import org.team1540.robot2024.util.auto.LocalADStarAK;
 import org.team1540.robot2024.util.vision.TimestampedVisionPose;
 
 import static org.team1540.robot2024.Constants.Drivetrain.*;
