+++ src/main/java/org/team1540/robot2024/util/auto/LocalADStarAK.java
@@ -1,4 +1,4 @@
-package org.team1540.robot2024.util;
+package org.team1540.robot2024.util.auto;
 
 import com.pathplanner.lib.path.GoalEndState;
 import com.pathplanner.lib.path.PathConstraints;
